expanded and finished asteroid example fixed some bugs when entities switch maps implemented missing asset functions misc documentation fixes minor clean ups added camera position offset option added exists check option per compile time config

+++ include/magique/ecs/Components.h
@@ -103,7 +103,7 @@ namespace magique
         [[nodiscard]] AnimationState getCurrentState() const;
 
         // Returns the current sprite count
-        [[nodiscard]] uint8_t getSpriteCount() const;
+        [[nodiscard]] uint16_t getSpriteCount() const;
 
     private:
         const EntityAnimation* entityAnimation; // Always valid
